refactor(cli): split checkpoint lineage verifier module
